+++ src/providers/postgres/qgspostgresfeatureiterator.cpp
@@ -81,7 +81,7 @@ QgsPostgresFeatureIterator::QgsPostgresFeatureIterator( QgsPostgresFeatureSource
   {
     QgsPostgresExpressionCompiler compiler = QgsPostgresExpressionCompiler( source );
 
-    if ( compiler.compile( request.filterExpression() ) == QgsPostgresExpressionCompiler::Complete )
+    if ( compiler.compile( request.filterExpression() ) == QgsSqlExpressionCompiler::Complete )
     {
       whereClause = QgsPostgresUtils::andWhereClauses( whereClause, compiler.result() );
       mExpressionCompiled = true;
